move operation class to operations/
