+++ lib/collection.js
@@ -60,8 +60,8 @@ const updateDocuments = require('./operations/collection_ops').updateDocuments;
 const updateMany = require('./operations/collection_ops').updateMany;
 const updateOne = require('./operations/collection_ops').updateOne;
 
-const InsertOneOperation = require('./operation_classes/insert_one');
-const executeOperationV2 = require('./execute_operation_v2');
+const InsertOneOperation = require('./operations/insert_one');
+const executeOperationV2 = require('./operations/execute_operation_v2');
 
 /**
  * @fileOverview The **Collection** class is an internal class that embodies a MongoDB collection
