+++ lib/operations/execute_operation_v2.js
