+++ lib/operations/insert_one.js
