+++ lib/operations/operation.js
